% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnosticLM}
\alias{diagnosticLM}
\title{a function for representing the residuals of a linear model}
\usage{
diagnosticLM(fitin)
}
\arguments{
\item{fitin}{is a list containing objects of class lm}
}
\value{
2 graphs in one page: a graph to check the homoscedasticity of residuals
        the second to check the normality of the residuals
}
\description{
a function for representing the residuals of a linear model
}
\examples{
\donttest{
 diagnosticLM(fitin=aListResult[[2]])
}
}
