% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingSplinesModel.R
\name{fitGSS}
\alias{fitGSS}
\title{a function to model curves using smoothing splines anova using \code{gss} library}
\usage{
fitGSS(datain, trait, loopId)
}
\arguments{
\item{datain}{input dataframe}

\item{trait}{character, trait of interest to model (example biovolume, PH ...)}

\item{loopId}{a column name that contains ident of Genotype-Scenario}
}
\value{
a list containing 2 objects
\describe{
\item{}{a list of each output of \code{\link[=ssanova]{ssanova}}}
\item{}{a dataframe of kullback-Leibler projection}
}
}
\description{
this function models each curve of genotype using smoothing
splines anova
}
\details{
the input dataframe must contain the following columns: the trait to model,
the ident of Genotype-Scenario, thermalTime, repetition columns
}
\examples{
\donttest{
 fm1<-fitGSS(datain=plant1,trait="biovolume")
}
}
\seealso{
\code{\link[=project.ssanova]{project.ssanova}}, \code{\link[=ssanova]{ssanova}}
}
