% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelisation.R
\name{fitLocfit}
\alias{fitLocfit}
\title{a function to model curve of a dataset using a local regression}
\usage{
fitLocfit(datain, trait, myylim, tt, reference)
}
\arguments{
\item{datain}{input dataframe}

\item{trait}{character, trait of interest to model (example biovolume, PH ...)}

\item{myylim}{numeric, max value of y-axis , varies according to trait}

\item{tt}{numeric vector of thermal times at which predictions are recovered}

\item{reference}{character, a column name indicating an unique identifiant of plant
(example: Ref == concatenation of experimentAlias-Line-Position-scenario)}
}
\value{
a dataframe of the predictions at tt thermal times for each plant of the input data set
and the graphics of the smoothing curves
}
\description{
this function models each curve of a dataset using a local regression
on biovolume, plant height or leaf area trait and extracts predictions at specific thermal
times
}
\details{
the input dataset must contain thermalTime column
}
\examples{
\donttest{
# Take a while...
 myThermalTimes<-c(24,30)
 resu<-fitLocfit(datain=plant1,trait="biovolume",myylim=700,tt=myThermalTimes,reference="Ref")
 str(resu)
}
}
