% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{fitSpATS}
\alias{fitSpATS}
\title{fitSpATS}
\usage{
fitSpATS(datain, trait, genotypeId, rowId, colId, typeModel = "anova",
  genotype.as.random = FALSE, nseg = c(14, 30), verbose)
}
\arguments{
\item{datain}{input dataframe of parameters}

\item{trait}{character, parameter of interest (ex: Biomass24)}

\item{genotypeId}{character, column name of genotype alias}

\item{rowId}{character, column name of the row id in the lattice (greenhouse or field)}

\item{colId}{character, column name of the column id in the lattice (greenhouse or field)}

\item{typeModel}{character, choice of the spatial P-Spline model, \code{anova} or \code{sap}}

\item{genotype.as.random}{logical. If TRUE, the genotype is included as random effect in the model. The default is FALSE.}

\item{nseg}{numerical vector of length 2 containing the number of segments for each marginal (strictly nseg - 1 is the
number of internal knots in the domain of the covariate).
Atomic values are also valid, being recycled. Default set to c(14,30)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
a SpATS object
}
\description{
a function to model curves using a spatial P-Spline modelling using \code{\link[=SpATS]{SpATS}} library
}
\details{
the input dataset must contain Position,Line,Ref,scenario,genotypeAlias columns
}
\examples{
\donttest{
 library(phisStatR)
 mydata<-plant4
 test<-fitSpATS(datain=mydata,trait="Biomass24",genotypeId="genotypeAlias",rowId="Line",
       colId="Position",typeModel="anova",genotype.as.random=FALSE,nseg=c(14,30),verbose)
}
}
\seealso{
\code{\link[=SpATS]{SpATS}}, \code{\link[=PSANOVA]{PSANOVA}} and \code{\link[=SAP]{SAP}}
}
