% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCleaning.R
\name{flagPointLocfit}
\alias{flagPointLocfit}
\title{flagPointLocfit}
\usage{
flagPointLocfit(datain, trait, xvar, loopID, locfit.h, threshold)
}
\arguments{
\item{datain}{input dataframe of parameters}

\item{trait}{character, parameter of interest (ex: plantHeight)}

\item{xvar}{character, time variable (ex: thermalTime)}

\item{loopID}{character, ID on which to make the loop}

\item{locfit.h}{numeric, the constant component of the smoothing parameter}

\item{threshold}{numeric, threshold to detect on the prediction interval}
}
\value{
a list
\describe{
\item{1}{prediction and detection of outlier on observed data}
\item{2}{prediction on regular abscissa data}
\item{3}{time courses with not enough point to be evaluated}
}
}
\description{
detection of outlier points in time courses
}
\examples{
\donttest{
# Example
}
}
