% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticModels.R
\name{fmGlobal}
\alias{fmGlobal}
\title{a function to modelize a trait in different trials taking into account the
position in the greenhouse (spatial correlation)}
\usage{
fmGlobal(datain, trait)
}
\arguments{
\item{datain}{input dataframe of parameters}

\item{trait}{character, trait of interest to model (example biovolume24, PH24 ...)}
}
\value{
a class asreml object
}
\description{
this function
}
\details{
This function needs in input a dataframe with the following description columns
\describe{
\item{Manip}{Id of Manip column}
\item{Genotype}{Genotype column}
\item{Line}{line column in the greenhouse}
\item{Position}{position column in the greenhouse}
\item{Repsce}{combination of Repetition and Scenario}
}
and require asreml library
}
\examples{
# not run
# library(asreml)
# df1<-fmGlobal(datain=,trait="biovolume24")
}
