% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSTmodel.R
\name{outlierCARBayesST}
\alias{outlierCARBayesST}
\title{outlierCARBayesST}
\usage{
outlierCARBayesST(modelin, datain, threshold, trait)
}
\arguments{
\item{modelin}{a CARBayesST object}

\item{datain}{input dataframe of parameters used in the CARBayesST modelling}

\item{threshold, }{numeric threshold for influence with standardized residuals}

\item{trait, }{character, trait of interest to model (example plantHeight...)}
}
\value{
a plot of fitted vs. residuals and a dataset with a column identifying the outlier points
}
\description{
a function to detect outlier point in time-course using a CARBayesST modelling
}
\examples{
\donttest{
 test<-fitCARBayesST(datain=mydata,xvar="thermalTime",trait="plantHeight",k=2,
     graphDist=TRUE,burnin=10,n.sample=110,
     formulaModel=as.formula(plantHeight~scenario+genotypeAlias),
     typeModel="anova",verbose=FALSE)
 test2<-outlierCARBayesST(modelin=test[[1]],datain=test[[2]],threshold=4,trait="plantHeight")
}
}
\seealso{
\code{\link[=CARBayesST]{CARBayesST}}, \code{\link[=ST.CARanova]{ST.CARanova}}
}
