% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{outlierCriteria}
\alias{outlierCriteria}
\title{a function for several outlier criteria}
\usage{
outlierCriteria(datain, typeD, residin, typeR, trait, resRawName,
  resStdName, threshold)
}
\arguments{
\item{datain}{input dataframe of parameters}

\item{typeD}{type of datain dataframe (1==wide, 2==long)}

\item{residin}{input dataframe of residuals}

\item{typeR}{type of residin dataframe (1==wide, 2==long)}

\item{trait}{character, trait of interest to model (example biovolume24, PH24 ...)}

\item{resRawName}{character, names of the raw residual in datain}

\item{resStdName}{character, names of the standardized residual in datain}

\item{threshold, }{numeric threshold for the normal quantile in raw criteria}
}
\value{
a dataframe with columns identifiying criteria used to detect outlier plants
with 1==plant OK - 0==plant KO to suppress
\describe{
\item{critraw}{raw criterion, critci: quartiles criterion}
\item{critinfl}{influence criterion with standardized residuals}
}
}
\description{
this function calculates outlier criteria on plants for each parameter
}
\details{
This function needs in input a dataframe with residuals extracted from
a mixed linear model (for instance using asreml or nlme libraries) and an another dataframe
with the estimated parameters (biovolume, plantHeight, leafArea etc...). Several criteria
will be calculated using different types of residuals. The 2 input dataframe must contain
the following columns names: "experimentAlias","Line" and "Position".
\describe{
\item{raw  and quartile criteria}{use raw residuals}
\item{influence criterion}{uses standardized residuals}
}
The function must be executed for each parameter of interest: biovolume, plantHeight and phy.
Each criteria will be used according with some rules:
\describe{
\item{Small plant}{biovolume and phy}
\item{Big plants}{biovolume and plantHeight}
}
}
\examples{
# Not run
# dt1<-outlierCriteria()

}
