% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant1}
\alias{plant1}
\title{Parameters of interest of plant over thermal time in an experiment}
\format{a data frame with XX rows and 14 variables:
\describe{
  \item{experimentAlias}{Id of an experiment}
  \item{Day}{date}
  \item{potAlias}{Id of the pot in the greenhouse}
  \item{scenario}{Scenarios of the experiment}
  \item{genotypeAlias}{Genotypes used in the experiment}
  \item{repetition}{Repetition}
  \item{Line}{Line position in the greenhouse}
  \item{Position}{Position in the line in the greenhouse}
  \item{thermalTime}{Thermal Time}
  \item{plantHeight}{plant height (mm)}
  \item{leafArea}{Leaf area}
  \item{biovolume}{biovolume (g)}
  \item{Repsce}{repetion-scenario}
  \item{Ref}{Reference of each plant in the greenhouse}
}}
\usage{
plant1
}
\description{
A dataset containing biovolume, plant height (PH) and leaf area (LA) over thermal time
for each plant of an experiment
}
\keyword{datasets}
