% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant2}
\alias{plant2}
\title{Visible leafs of plant over thermal time in an experiment}
\format{a data frame with XX rows and 10 variables:
\describe{
  \item{experimentAlias}{Id of an experiment}
  \item{F_visible}{Number of leafs}
  \item{DAS}{DAS}
  \item{Line}{Line position in the greenhouse}
  \item{Position}{Position in the line in the greenhouse}
  \item{scenario}{Scenarios of the experiment}
  \item{genotypeAlias}{Genotypes used in the experiment}
  \item{Genetic}{Genetic}
  \item{thermalTime}{Thermal Time}
  \item{Ref}{Reference of each plant in the greenhouse}
}}
\usage{
plant2
}
\description{
A dataset containing the number of leafs over thermal time for each plant
of an experiment
}
\keyword{datasets}
