% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant4}
\alias{plant4}
\title{Parameters of interest of plants in an experiment for outlier detection purpose}
\format{a data frame with 1680 rows and 11 variables:
\describe{
  \item{Line}{Line position in the greenhouse}
  \item{Position}{Position in the line in the greenhouse}
  \item{Ref}{Reference of each plant in the greenhouse}
  \item{potAlias}{Id of the pot in the greenhouse}
  \item{scenario}{Scenarios of the experiment}
  \item{genotypeAlias}{Genotypes used in the experiment}
  \item{repetition}{Repetition}
  \item{Repsce}{repetion-scenario}
  \item{Biomass24}{biomass at 24 TT day (g)}
  \item{PH24}{plant height at 24 TT day (mm)}
  \item{Phy}{Phyllocron}
}}
\usage{
plant4
}
\description{
A dataset containing biovolume (biovolume24), plant height (PH24) and phyllocron (phy)
for each plant of an experiment for outlier detection purpose
}
\keyword{datasets}
