% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalFunctions.R
\name{plotCARBayesST}
\alias{plotCARBayesST}
\title{a function for representing a CARBayesST result}
\usage{
plotCARBayesST(datain, outlierin, myselect, trait, xvar)
}
\arguments{
\item{datain}{the input dataframe used in \code{\link[=fitCARBayesST]{fitCARBayesST}} to detect outlier points in time courses}

\item{outlierin}{the input dataframe identifying the outlier points in the time courses, output of the \code{\link[=outlierCARBayesST]{outlierCARBayesST}}}

\item{myselect}{a numeric vector for facet graph}

\item{trait}{a variable to explore}

\item{xvar}{character, time variable (ex: thermalTime)}
}
\value{
a ggplot2 graph object
}
\description{
a function for representing a CARBayesST result
}
\details{
the input dataset must contain scenario,genotypeAlias and Repsce (combination of repetition and scenario) columns
}
\examples{
# not run
}
