% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCleaning.R
\name{plotFlagPoint}
\alias{plotFlagPoint}
\title{plotFlagPoint}
\usage{
plotFlagPoint(smoothin, loopID, myselect)
}
\arguments{
\item{smoothin}{the 1st element of a flagPointLocfit() list result. smoothin contains the prediction on observed data and the detected outlier}

\item{loopID}{character, name of the column of the time courses' ID}

\item{myselect}{character, a vector of ID}
}
\value{
a graph
}
\description{
plot detection of outlier points in time courses
}
\examples{
\donttest{
# Example
}
}
