% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalFunctions.R
\name{plotGSS}
\alias{plotGSS}
\title{a function for representing a smoothing spline anova result}
\usage{
plotGSS(datain, modelin, trait, myvec, lgrid)
}
\arguments{
\item{datain}{a dataframe}

\item{modelin}{a list of output of ssanova()}

\item{trait}{a variable to explore}

\item{myvec}{a numeric vector for facet graph}

\item{lgrid}{length of the regular grid for ssanova prediction}
}
\value{
a ggplot2 graph object
}
\description{
produce a graphic of Smoothing splines anova fitting
}
\details{
the dataframe in input must contain Genosce, repetition and thermalTime columns!
}
\examples{
# Not run

}
