% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSTmodel.R
\name{printCARBayesST}
\alias{printCARBayesST}
\title{printCARBayesST}
\usage{
printCARBayesST(modelin)
}
\arguments{
\item{modelin}{a CARBayesST object to describe}
}
\value{
a summary of a CARBayesST object (spatio-temporal bayesian modelling)
}
\description{
a function for CARBayesST analysis description
}
\examples{
\donttest{
 test<-fitCARBayesST(datain=mydata,xvar="thermalTime",trait="plantHeight",k=2,
     graphDist=TRUE,burnin=10,n.sample=110,
     formulaModel=as.formula(plantHeight~scenario+genotypeAlias),
     typeModel="anova",verbose=FALSE)
 printCARBayesST(modelin=test[[1]])
}
}
\seealso{
\code{\link[=CARBayesST]{CARBayesST}}, \code{\link[=ST.CARanova]{ST.CARanova}}
}
