% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printExperiment.R
\name{printExperiment}
\alias{printExperiment}
\title{a function for experiment description}
\usage{
printExperiment(datain, manip)
}
\arguments{
\item{datain}{a dataframe to describe}

\item{manip}{an identifiant of an experiment (example ZA16)}
}
\value{
a description
}
\description{
a function for experiment description
}
\details{
the input dataframe must contain the following columns: experimentAlias,
genotypeAlias, scenario, Repsce and Ref
}
\examples{
\donttest{
# Not run
 printExperiment(datain=toto,manip="ZB12")
}
}
