% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingSplinesModel.R
\name{printGSS}
\alias{printGSS}
\title{a function for gss analysis description}
\usage{
printGSS(object, threshold)
}
\arguments{
\item{object}{a dataframe to describe from fitGSS() function}

\item{threshold}{numeric, a threshold for Kullback-Leibler projection}
}
\value{
a description
}
\description{
a function for gss analysis description
}
\details{
the input object is the 2nd element of a fitGSS result, a dataframe with the Kullback-Leibler projection
         colnames of this dataframe are Genosce, ratio, kl, check
}
\examples{
\donttest{
 fm1<-fitGSS(datain=plant1,trait="biovolume")
 printGSS(object=fm1,threshold=0.05)
}
}
\seealso{
\code{\link[=ssanova]{ssanova}}
}
