% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalTime.R
\name{thermalTime}
\alias{thermalTime}
\title{a function to calculate thermal time}
\usage{
thermalTime(datain, inSpecie, method, inDateS = NULL, inDateE = NULL,
  inTemp = NULL)
}
\arguments{
\item{datain}{input dataframe of meteo data from phis web service}

\item{inSpecie}{character, studied specie}

\item{method}{character, a method of thermal time's calculation ("parent","baseline")}

\item{inDateS}{a date of sowing or thinning etc... start event ("YYYY-MM-DD")}

\item{inDateE}{a date of harvesting ("YYYY-MM-DD")}

\item{inTemp}{numeric, a baseline temperature for baseline's method}
}
\value{
a dataframe
}
\description{
this function calculates the thermal time for an experiment
             according to several methods
}
\details{
Parent et. al. (2010). Modelling temperature-compensated physiological rates,
   based on the co-ordination of responses to temperature of developmental processes.
   Journal of Experimental Botany. 61 (8):2057-2069

   if the Parent's model is chosen inTemp must be NULL

   The input dataframe is extracted from phis web service (getEnvironment function)
   and is structured as follow: date, value, sensor, codeVariable and facility
}
\examples{
\donttest{
# Example for the model of Parent 2010
library(phisWSClientR)
# initializeClientConnection and getEnvironment are functions of the phisWSClientR library
initializeClientConnection(apiID="ws_public")
token <- getToken("guestphis@supagro.inra.fr","guestphis",verbose=FALSE)
tpCount<-getEnvironment(token = token$data ,
              experimentURI="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
              facility="http://www.phenome-fppn.fr/m3p/es2",
              variables="leaf temperature_thermocouple sensor_degree celsius",
              verbose=FALSE)$totalCount
myMeteo<-getEnvironment(token = token$data,
              experimentURI="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
              facility="http://www.phenome-fppn.fr/m3p/es2",
              variables="leaf temperature_thermocouple sensor_degree celsius",
              pageSize=tpCount,verbose=FALSE)$data
test<-thermalTime(datain=myMeteo,inSpecie="maize",method="parent",inDateS="2012-01-10",
                  inDateE="2012-03-10",inTemp=NULL)
}
}
