% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printExperiment.R
\name{printExperiment}
\alias{printExperiment}
\title{a function for experiment description}
\usage{
printExperiment(datain)
}
\arguments{
\item{datain}{a dataframe to describe containing a unique experiment}
}
\value{
a description
}
\description{
a function for experiment description
}
\details{
the input dataframe must contain the following columns: experimentAlias,
Genotype, scenario, coordinates and pot.
}
\examples{
\donttest{
  printExperiment(datain=plant3)
}
}
